"Add MarkdownV2 parsing to bot messages"
This commit adds MarkdownV2 parsing to the bot's response messages. Previously, the bot would send plain text messages which doesn't allow for formatting such as bold or italic text. With the addition of MarkdownV2 parsing, the bot's messages can now include more complex text formatting. This change aims to improve the readability and presentation of the bot's messages.